test2

+++ example.c
@@ -157,7 +157,7 @@ bool module_init(const module_config_t *cfg)
 
     int a = 5;
     unsigned int rk = 0;
-    rk = ((int)a);
+    rk = a;
     
     prv_log(LOG_INFO, "module_init: cap=%u level=%d", s_ring.capacity, (int)s_log_level);
     return true;
